MODEXPW-596. Use the same construction total.getNumber().doubleValue() as in original PR

+++ src/main/java/org/folio/dew/batch/acquisitions/edifact/CompositePOLineConverter.java
@@ -500,6 +500,6 @@ public class CompositePOLineConverter {
       total = total.add(ePrice);
     }
 
-    return total.getNumber();
+    return total.getNumber().doubleValue();
   }
 }
